Moved resources rendering into client side component
