+++ src/app/(resources)/media/page.tsx
@@ -1,13 +1,6 @@
-import { ResourcesTitle } from "@/app/components/resourcesTitle";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { faYoutube, faSpotify, IconDefinition } from "@fortawesome/free-brands-svg-icons";
-import ScrollToHash from "@/utils/scrollToHash";
 import '@/styles/resources.scss'
-import Feedback from "@/app/components/feedback";
-import Image from "next/image";
-import Sidebar from "@/app/components/sidebar";
-import Navbar from "@/app/components/navbar";
-import {getResources} from "@/app/api/resources-api";
+import {getResources} from "@/app/api/api";
+import { LoadMedia } from "@/app/components/resources/loadMedia";
 
 type Media = {
     id: number,
@@ -36,45 +29,9 @@ export default async function Media() {
         media = {}; // Fallback to empty data or provide some default data
     }
 
-    const platforms_icons: { [key: string]: IconDefinition } = {
-        'youtube': faYoutube,
-        'spotify': faSpotify
-    }
-
     return (
         <>
-            <Navbar resources={media} />
-            <div className="resources-wrapper">
-                <div className="resources-container">
-                    <ScrollToHash />
-                    <ResourcesTitle title="Media Library 💾" description="Teaching videos, vlogs, podcasts, gaming and more!" />
-                    <Feedback
-                        title="Share your favorite Youtube channel or podcast 📀"
-                        categories={Object.keys(media)} />
-                    <hr className="resources-title-separator" />
-                    <div className="list-container media">
-                        {Object.keys(media).map((category, i) => (
-                            <section key={i} className={`${media[category][0].slug}-container`}>
-                                <a id={`${media[category][0].slug}-id`} className="anchor"></a>
-                                <h2 className="category-title">{category}</h2>
-                                <ul key={i} className="list-item-container media">
-                                    {media[category].map((item: Media) => (
-                                        <li key={item.uuid} className={`item-container ${item.slug} media`}>
-                                            <Image src={item.image} alt={item.name} width={60} height={60} className="item-picture media" />
-                                            <div className="item-infos">
-                                                <a href={item.link} className="item media" target="_blank">{item.name} <FontAwesomeIcon className="icon" icon={platforms_icons[item.platform]} width={20} /></a>
-                                                {item.description}
-                                            </div>
-                                            <Image src={item.image} alt={item.name} width={100} height={100} className="item-picture-bg" />
-                                        </li>
-                                    ))}
-                                </ul>
-                            </section>
-                        ))}
-                    </div>
-                </div>
-                <Sidebar resources={media} />
-            </div>
+            <LoadMedia media={media}  />
         </>
     )
 }
