+++ src/app/(resources)/resources/page.tsx
@@ -1,15 +1,6 @@
-import { ResourcesTitle } from "@/app/components/resourcesTitle"
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { faFlagCheckered, faBook, faSpellCheck, faLanguage, faBookOpen, faBookmark, IconDefinition } from "@fortawesome/free-solid-svg-icons";
-import ScrollToHash from "@/utils/scrollToHash";
 import '@/styles/resources.scss'
-import Feedback from "@/app/components/feedback";
-import Sidebar from "@/app/components/sidebar";
-import Navbar from "@/app/components/navbar";
-import { getFavorite, getResources } from "@/app/api/api";
-import AddToFavorite from "@/app/components/addFavorite";
-import { cookies } from "next/headers";
-
+import { getResources } from "@/app/api/api";
+import { ShowResources } from "@/app/components/resources/loadResources";
 
 type Resource = {
     id: number,
@@ -36,51 +27,9 @@ export default async function Resources() {
         resources = {}; // Fallback to empty data or provide some default data
     }
 
-    const category_icons: { [key: string]: IconDefinition } = {
-        'beginner': faFlagCheckered,
-        'dictionaries': faBook,
-        'grammar': faSpellCheck,
-        'vocabulary': faLanguage,
-        'reading': faBookOpen,
-        'miscellaneous': faBookmark
-    }
-
     return (
         <>
-            <Navbar resources={resources} />
-            <div className="resources-wrapper">
-                <div className="resources-container">
-                    <ScrollToHash />
-                    <ResourcesTitle title="Study Resources 📖" description="Dictionaries, grammar guides, vocabulary insights, reading materials and other useful resources for studying!" />
-                    <Feedback
-                        title="Share your favorite Japanese learning tool 📖"
-                        categories={Object.keys(resources)} />
-                    <hr className="resources-title-separator" />
-                    <div className="list-container">
-                        {Object.keys(resources).map((category, i) => (
-                            <section key={i} className={`${resources[category][0].slug}-container`}>
-                                <a id={`${resources[category][0].slug}-id`} className="anchor"></a>
-                                <h2 className="category-title">{category}</h2>
-                                <ul key={i} className="list-item-container">
-                                    {resources[category].map((item: Resource) => (
-                                        <li key={item.uuid} className={`item-container ${item.slug}`}>
-                                            <FontAwesomeIcon className="item-icons" icon={category_icons[item.slug]} height={20} />
-                                            <a href={item.link} className="item" target="_blank">{item.name}</a>
-                                            {` - ${item.description}`}
-                                            <div className="add-to-favorite" style={{ display: 'inline' }}>
-                                                <AddToFavorite
-                                                    id={item.uuid}
-                                                    type={'resources'}/>
-                                            </div>
-                                        </li>
-                                    ))}
-                                </ul>
-                            </section>
-                        ))}
-                    </div>
-                </div>
-                <Sidebar resources={resources} />
-            </div>
+            <ShowResources resources={resources}/>
         </>
     )
 }
